[#10] Added basic tests for view and moved main database fixture to conftest.py.
